+++ datalad/interface/tests/test_rerun.py
@@ -67,9 +67,9 @@ from datalad.tests.utils import (
     assert_not_in,
     swallow_logs,
     swallow_outputs,
+    known_failure_appveyor,
     known_failure_windows,
     slow,
-    SkipTest,
 )
 
 
@@ -581,6 +581,9 @@ def test_rerun_script(path):
 
 
 @slow  # ~10s
+@known_failure_appveyor
+# ^ Issue only happens on appveyor, Python itself implodes. Cannot be
+#   reproduced on a real win7 box
 @with_tree(tree={"input.dat": "input",
                  "extra-input.dat": "extra input",
                  "s0": {"s1_0": {"s2": {"a.dat": "a",
@@ -590,12 +593,6 @@ def test_rerun_script(path):
                         "ss": {"e.dat": "e"}}})
 @with_tempfile(mkdir=True)
 def test_run_inputs_outputs(src, path):
-    if 'APPVEYOR' in os.environ:
-        # issue only happens on appveyor, Python itself implodes
-        # cannot be reproduced on a real win7 box
-        raise SkipTest(
-            'test causes appveyor (only) to crash, reason unknown')
-
     for subds in [("s0", "s1_0", "s2"),
                   ("s0", "s1_1", "s2"),
                   ("s0", "s1_0"),
@@ -758,6 +755,49 @@ def test_run_inputs_no_annex_repo(path):
     ds.run("cd .> dummy", inputs=["*"])
     ok_exists(op.join(ds.path, "dummy"))
     ds.rerun()
+
+
+@known_failure_windows
+@with_tree(tree={"to_modify": "to_modify"})
+def test_rerun_explicit(path):
+    ds = Dataset(path).create(force=True)
+
+    ds.run("echo o >> foo", explicit=True, outputs=["foo"])
+    with open(op.join(ds.path, "foo")) as ifh:
+        orig_content = ifh.read()
+        orig_head = ds.repo.get_hexsha()
+
+    # Explicit rerun is allowed in a dirty tree.
+    ok_(ds.repo.dirty)
+    ds.rerun(explicit=True)
+    eq_(orig_head, ds.repo.get_hexsha("master~1"))
+    with open(op.join(ds.path, "foo")) as ifh:
+        eq_(orig_content * 2, ifh.read())
+
+    # --since also works.
+    ds.rerun(since="", explicit=True)
+    eq_(orig_head,
+        # Added two rerun commits.
+        ds.repo.get_hexsha("master~3"))
+
+    # With just untracked changes, we can rerun with --onto.
+    ds.rerun(since="", onto="", explicit=True)
+    eq_(ds.repo.get_hexsha(orig_head + "^"),
+        # Reran the four run commits from above on the initial base.
+        ds.repo.get_hexsha("HEAD~4"))
+
+    # But checking out a new HEAD can fail when there are modifications.
+    ds.repo.checkout("master")
+    ok_(ds.repo.dirty)
+    ds.repo.add(["to_modify"], git=True)
+    ds.save()
+    assert_false(ds.repo.dirty)
+    with open(op.join(ds.path, "to_modify"), "a") as ofh:
+        ofh.write("more")
+    ok_(ds.repo.dirty)
+
+    with assert_raises(CommandError):
+        ds.rerun(onto="", since="", explicit=True)
 
 
 @with_tree(tree={"a.in": "a", "b.in": "b", "c.out": "c",