+++ CHANGELOG.md
@@ -19,11 +19,27 @@ bet we will fix some bugs and make a world even a better place.
 
 ### Fixes
 
-?
+- Our tests needed various adjustments to keep up with upstream
+  changes in Travis and Git. ([#3479][]) ([#3492][]) ([#3493][])
+
+- `AnnexRepo.is_special_annex_remote` was too selective in what it
+  considered to be a special remote.  ([#3499][])
+
+- We now provide information about unexpected output when git-annex is
+  called with `--json`.  ([#3516][])
 
 ### Enhancements and new features
 
-?
+- For calls to git and git-annex, we disable automatic garbage
+  collection due to past issues with GitPython's state becoming stale,
+  but doing so results in a larger .git/objects/ directory that isn't
+  cleaned up until garbage collection is triggered outside of DataLad.
+  Tests with the latest GitPython didn't reveal any state issues, so
+  we've re-enabled automatic garbage collection.  ([#3458][])
+
+- [rerun][] learned an `--explicit` flag, which it relays to its calls
+  to [run][[]].  This makes it possible to call `rerun` in a dirty
+  working tree ([#3498][]).
 
 ## 0.12.0rc4 (May 15, 2019) -- the revolution is over
 
@@ -783,8 +799,10 @@ A number of fixes did not make it into the 0.9.x series:
   included in the output by passing `--senstive=some` or `--senstive=all`.
 - Reduced startup latency by only importing commands necessary for a particular
   command line call.
-- `datalad create -d <parent> --nosave` now registers subdatasets, when possible.
-- `datalad run` now provides a way for the caller to save the result when a
+- [create]:
+  - `-d <parent> --nosave` now registers subdatasets, when possible.
+  - `--fake-dates` configures dataset to use fake-dates
+- [run] now provides a way for the caller to save the result when a
   command has a non-zero exit status.
 - `datalad rerun` now has a `--script` option that can be used to extract
   previous commands into a file.
@@ -1338,6 +1356,7 @@ publishing
 [clone]: http://datalad.readthedocs.io/en/latest/generated/man/datalad-clone.html
 [configuration]: http://docs.datalad.org/en/latest/config.html
 [copy_to]: http://docs.datalad.org/en/latest/_modules/datalad/support/annexrepo.html?highlight=%22copy_to%22
+[create]: http://datalad.readthedocs.io/en/latest/generated/man/datalad-create.html
 [create-sibling-github]: http://datalad.readthedocs.io/en/latest/generated/man/datalad-create-sibling-github.html
 [create-sibling]: http://datalad.readthedocs.io/en/latest/generated/man/datalad-create-sibling.html
 [datalad]: http://docs.datalad.org/en/latest/generated/man/datalad.html
@@ -1546,3 +1565,10 @@ publishing
 [#3425]: https://github.com/datalad/datalad/issues/3425
 [#3439]: https://github.com/datalad/datalad/issues/3439
 [#3440]: https://github.com/datalad/datalad/issues/3440
+[#3458]: https://github.com/datalad/datalad/issues/3458
+[#3479]: https://github.com/datalad/datalad/issues/3479
+[#3492]: https://github.com/datalad/datalad/issues/3492
+[#3493]: https://github.com/datalad/datalad/issues/3493
+[#3498]: https://github.com/datalad/datalad/issues/3498
+[#3499]: https://github.com/datalad/datalad/issues/3499
+[#3516]: https://github.com/datalad/datalad/issues/3516